requirements.txt
apscheduler>=3.10,<4.0
python-binance>=1.0.19,<2.0
python-dotenv>=1.0,<2.0
